Fix game persistence for starting a new game

+++ src/main/java/de/htwberlin/casino/blackjack/application/port/in/playGame/StartGameCommand.java
@@ -1,3 +1,3 @@
 package de.htwberlin.casino.blackjack.application.port.in.playGame;
 
-public record StartGameCommand(String userId, int bet) {}
+public record StartGameCommand(String userId, double bet) {}
